requirements.txt
python-dotenv==1.1.1
coinbase-advanced-py==1.8.2